Move type serializer tests
